Add files via upload
The previous version is now amended with the true values for the precision of private information and the real policy of the Bank of Russia, which are now 1/4.9 and -0.1 respectively. We had also revised our estimate for the strategic complementarity component, which is now 0.4.
Fortunately, it made our results even more compelling. See the new version in the Julia code!